Spending pages improvements (#19)
- Adds initial localization to Requests Index.
- Adds initial localization to Request Details.
- Improve user navigation by removing Edit buttons from Request Index.
- Prevent deletion of Requests except for just created ones.
- Only Administrators can Complete Requests.

+++ ConselvaBudget/Areas/Spending/Pages/Requests/Delete.cshtml.cs
@@ -42,7 +42,9 @@ namespace ConselvaBudget.Areas.Spending.Pages.Requests
 
             var expensesRequest = await _context.SpendingRequests.FindAsync(id);
 
-            if (expensesRequest != null)
+            // Only allow deletions of just created Requests
+
+            if (expensesRequest != null && expensesRequest.Status == RequestStatus.Created)
             {
                 ExpensesRequest = expensesRequest;
                 _context.SpendingRequests.Remove(ExpensesRequest);
